Arreglo float producto

+++ app/migrations/0003_product.py
@@ -16,7 +16,7 @@ class Migration(migrations.Migration):
                 ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
                 ('name', models.CharField(max_length=100)),
                 ('type', models.CharField(max_length=100)),
-                ('price', models.IntegerField()),
+                ('price', models.FloatField()),
             ],
         ),
     ]
